4D Printing from CFRP Composite to Make the Letter "S"
This report discusses the manufacturing of a Carbon-Fibre Reinforced Polymer (CFRP) structure
in the shape of the letter "S". The method adopted for this project is the 4D printing of composite
materials. The structure consisted of 5 layers in total; 1 24-inch-long 0° layer (spine) and 4 90°
layers of 13 inches each. The 90° layers are placed in 2 with each pile facing different directions and
on opposite ends of the spine.
The final product produced in this experiment was found to adhere with the project guidelines and
achieve the design goals that were set beforehand. The structure manufactured accurately portraits
the letter "S". It has also been tested against buckling and found that it can support its own weight
without distorting the shape too much.
The letter’s dimensions are 45.2 cm in height, 11.3 cm in width, and 7.7 cm in depth.